Fixed a problem with the module name
